+++ src/server/game/AI/CoreAI/PetAI.cpp
@@ -16,6 +16,7 @@
  */
 
 #include "PetAI.h"
+#include "CharmInfo.h"
 #include "Creature.h"
 #include "Errors.h"
 #include "Group.h"
